saves data and such

+++ FunkyTime.py
@@ -204,6 +204,18 @@ class Funky_GUI(wx.Frame):
         data = None
         while data is None:
             data,ctype = sr.poll_server(download_hash)
+        if ctype == 'audio/mpeg':
+            fileout = open('music/'+open_meta_data['title']+'.mp3', 'wb')
+            fileout.write(data)
+        if ctype == 'audio/m4a':
+            fileout = open('music/'+open_meta_data['title']+'.m4a', 'wb')
+            fileout.write(data)
+        if ctype == 'audio/x-flac':
+            fileout = open('music/'+open_meta_data['title']+'.flac', 'wb')
+            fileout.write(data)
+        if ctype == 'audio/x-wav':
+            fileout = open('music/'+open_meta_data['title']+'.wav', 'wb')
+            fileout.write(data)
 
     def onBrowse(self, event):
         """
@@ -223,6 +235,7 @@ class Funky_GUI(wx.Frame):
     def delete_file(self, event):
         files_deleted = self.playlist.removeSelected()
         for dfile in files_deleted:
+            if dfile.split('/')[-2] != 'music': continue
             os.remove(dfile)
 
     def add_file(self, event):
